Fixes issue #1 when running installation again failed

+++ fonts/install.sh
@@ -2,10 +2,19 @@
 echo "Installing Fonts"
 
 # Install Powerline fonts
-
-git clone https://github.com/powerline/fonts.git "$DOTFILES_ROOT/fonts/powerline-fonts"
-. "$DOTFILES_ROOT/fonts/powerline-fonts/install.sh"
-rm -rf "$DOTFILES_ROOT/fonts/powerline-fonts"
+powerline_folder="$DOTFILES_ROOT/fonts/powerline-fonts"
+if [ ! -d "$powerline_folder" ]; then
+    git clone https://github.com/powerline/fonts.git "$DOTFILES_ROOT/fonts/powerline-fonts"
+    . "$DOTFILES_ROOT/fonts/powerline-fonts/install.sh"
+    rm -rf "$DOTFILES_ROOT/fonts/powerline-fonts"
+fi
 
 # Installing Solarized 
-git clone git://github.com/altercation/solarized.git "$DOTFILES_ROOT/zsh/solarized"
+solarized_folder="$DOTFILES_ROOT/zsh/solarized"
+if [ ! -d "$solarized_folder" ]; then
+    echo "Cloning Solarzied into $solarized_folder"
+    git clone git://github.com/altercation/solarized.git "$DOTFILES_ROOT/zsh/solarized"
+else
+    echo "Pulling Solarzied"
+    git pull
+fi
